modularize dataset imports, update Next.js type references, and add CLAUDE.md documentation
